Renamed libraries, but using original function names.
This way one knows if he is using the old or STM library by looking at
the include line, while the sketch should compile file as the functions
are named the same. I believe this is the best compromise.
