+++ STM32F1/libraries/Adafruit_GFX_AS/Adafruit_GFX_AS.cpp
@@ -31,7 +31,7 @@ ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 POSSIBILITY OF SUCH DAMAGE.
 */
 
-#include "Adafruit_GFX.h"
+#include "Adafruit_GFX_AS.h"
 
 #ifdef LOAD_GLCD
   #include "glcdfont.c"
