+++ STM32F1/libraries/Adafruit_ILI9341_STM/README.txt
