+++ STM32F1/libraries/Adafruit_ILI9341_STM/examples/TFT_Clock_Digital_ILI9341/TFT_Clock_Digital_ILI9341.ino
@@ -37,8 +37,8 @@ code	color
 #define dc   10
 #define rst  9  // you can also connect this to the Arduino reset
 
-#include <Adafruit_GFX.h>    // Core graphics library
-#include <Adafruit_ILI9341.h> // Hardware-specific library
+#include <Adafruit_GFX_AS.h>    // Core graphics library, with extra fonts.
+#include <Adafruit_ILI9341_STM.h> // STM32 DMA Hardware-specific library
 #include <SPI.h>
 
 #define ILI9341_GREY 0x5AEB
@@ -55,7 +55,7 @@ unsigned int colour = 0;
 
 void setup(void) {
   tft.begin();
-  tft.setRotation(4);
+  tft.setRotation(2);
   tft.fillScreen(ILI9341_BLACK);
 
   tft.setTextSize(1);
