+++ STM32F1/libraries/Adafruit_ILI9341_STM/examples/TFT_Rainbow_ILI9341/TFT_Rainbow_ILI9341.ino
@@ -17,9 +17,10 @@
 //#define mosi 11  // Don't change
 #define cs   8
 #define dc   10
-#define rst  9 // you can also connect this to the Arduino reset
-#include <Adafruit_GFX.h>    // Core graphics library
-#include <Adafruit_ILI9341.h> // Hardware-specific library
+#define rst  9  // you can also connect this to the Arduino reset
+
+#include <Adafruit_GFX_AS.h>    // Core graphics library, with extra fonts.
+#include <Adafruit_ILI9341_STM.h> // STM32 DMA Hardware-specific library
 #include <SPI.h>
 
 Adafruit_ILI9341 tft = Adafruit_ILI9341(cs, dc, rst);       // Invoke custom library
